Rename the file to numberFormat
