--- src/utils/applyThousandSeparator.ts
@@ -1,5 +1,0 @@
-export function applyThousandSeparator(str: string) {
-  const thousandReg = /\B(?=(\d{3})+(?!\d))/g;
-
-  return str.replace(thousandReg, ',');
-}
